requirements.txt
python-telegram-bot==20.8
yoomoney==0.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0
loguru==0.7.2
pydantic==2.5.2